requirements.txt
# HIT140 Assessment 3 - Bat Seasonal Behavior Analysis
# Data Science Dependencies

# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter Environment
jupyter>=1.0.0
ipython>=8.0.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=0.20.0
click>=8.0.0

# Development and Testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0